Tests/LibWeb/Text/expected/navigation/canceled-push-then-intercepted-traverse.txt

navigate push destination=?canceled cancelable=true
navigate traverse destination=?start
traverse handler
final location=?start index=0
